+++ lib/src/models/currency_info.dart
@@ -3,11 +3,13 @@ class CurrencyInfo {
   final String name;
   final String symbol;
   final String flag;
+  final bool isCrypto;
 
   const CurrencyInfo({
     required this.code,
     required this.name,
     required this.symbol,
     required this.flag,
+    this.isCrypto = false,
   });
 }